+++ src/util/intset.rs
@@ -1,14 +1,14 @@
 pub struct IntSet {
-    min: u32,
-    max: u32,
+    min: usize,
+    max: usize,
     items: Vec<usize>
 }
 
 impl IntSet {
-    pub fn new(min: u32, max: u32) -> IntSet {
+    pub fn new(min: usize, max: usize) -> IntSet {
         let num_items = max - min + 1;
         let mut items: Vec<usize> = vec!();
-        items.resize(((num_items + usize::BITS - 1) / usize::BITS) as usize, 0);
+        items.resize((num_items + usize::BITS as usize - 1) / usize::BITS as usize, 0);
         IntSet {
             min,
             max,
@@ -16,17 +16,17 @@ impl IntSet {
         }
     }
 
-    pub fn add(&mut self, mut item: u32) {
+    pub fn add(&mut self, mut item: usize) {
         item -= self.min;
-        let item_index = item / usize::BITS;
-        let bit_offset = item % usize::BITS;
+        let item_index = item / usize::BITS as usize;
+        let bit_offset = item % usize::BITS as usize;
         self.items[item_index as usize] |= 1 << bit_offset;
     }
 
-    pub fn remove(&mut self, mut item: u32) {
+    pub fn remove(&mut self, mut item: usize) {
         item -= self.min;
-        let item_index = item / usize::BITS;
-        let bit_offset = item % usize::BITS;
+        let item_index = item / usize::BITS as usize;
+        let bit_offset = item % usize::BITS as usize;
         self.items[item_index as usize] &= !(1 << bit_offset);
     }
 
@@ -34,10 +34,10 @@ impl IntSet {
         self.items.fill(0);
     }
 
-    pub fn contains(&self, mut item: u32) -> bool {
+    pub fn contains(&self, mut item: usize) -> bool {
         item -= self.min;
-        let item_index = item / usize::BITS;
-        let bit_offset = item % usize::BITS;
+        let item_index = item / usize::BITS as usize;
+        let bit_offset = item % usize::BITS as usize;
         (self.items[item_index as usize] & (1 << bit_offset)) != 0
     }
 
@@ -50,11 +50,11 @@ impl IntSet {
         }
     }
 
-    pub fn get_min(&self) -> u32 {
+    pub fn get_min(&self) -> usize {
         self.min
     }
 
-    pub fn get_max(&self) -> u32 {
+    pub fn get_max(&self) -> usize {
         self.max
     }
 }
